Closer to working with 25, removes warnings as errors setting on scalac.
