+++ app/controllers/Homepage.scala
@@ -10,7 +10,7 @@ import play.api.mvc._
 import SessionKeys.SupplierTrackingCode
 import utils.Tracking.internalCampaignCode
 
-object Homepage extends Controller {
+class Homepage extends Controller {
 
   def index = NoCacheAction { implicit request =>
     val countryGroup = request.getFastlyCountryGroup.getOrElse(CountryGroup.UK)
