+++ app/controllers/OAuth.scala
@@ -1,20 +1,23 @@
 package controllers
 
+import javax.inject.Inject
+
 import com.gu.googleauth.GoogleAuthFilters.LOGIN_ORIGIN_KEY
-import actions.CommonActions.NoCacheAction
+ import actions.OAuthActions
 import com.gu.googleauth.{GoogleAuth, UserIdentity}
 import configuration.Config
 import model.FlashMessage
 import play.api.Play.current
 import play.api.libs.json.Json
 import play.api.mvc._
-
 import play.api.libs.concurrent.Execution.Implicits.defaultContext
+import play.api.libs.ws.WSClient
+
 import scala.concurrent.Future
 
-object OAuth extends Controller {
+class OAuth @Inject()(val wsClient: WSClient) extends Controller with OAuthActions{
   val ANTI_FORGERY_KEY = "antiForgeryToken"
-
+  implicit val iWsClient = wsClient
   def login = NoCacheAction { request =>
     val flashMsgOpt = request.flash.get("error").map(FlashMessage.error)
     Ok(views.html.staff.unauthorised(flashMsgOpt))
@@ -25,6 +28,7 @@ object OAuth extends Controller {
    * Redirect to Google with anti forgery token (that we keep in session storage - note that flashing is NOT secure)
    */
   def loginAction = Action.async { implicit request =>
+    implicit val iWsClient = wsClient
     val antiForgeryToken = GoogleAuth.generateAntiForgeryToken()
     GoogleAuth.redirectToGoogle(Config.googleAuthConfig, antiForgeryToken)
       .map(_.withSession(request.session + (ANTI_FORGERY_KEY -> antiForgeryToken)))
