+++ app/controllers/PatternLibrary.scala
@@ -1,9 +1,9 @@
 package controllers
 
+import actions.CommonActions
 import play.api.mvc.Controller
-import actions.CommonActions.NoCacheAction
 
-object PatternLibrary extends Controller {
+object PatternLibrary extends Controller with CommonActions{
 
   def patterns = NoCacheAction { implicit request =>
     Ok(views.html.patterns.patterns())
