+++ app/filters/HandleXFrameOptionsOverrideHeader.scala
@@ -7,13 +7,14 @@ import play.api.libs.concurrent.Execution.Implicits.defaultContext
 import play.api.mvc._
 
 import scala.concurrent.Future
-
-class HandleXFrameOptionsOverrideHeader @Inject()(implicit  val mat: Materializer) extends Filter {
+object HandleXFrameOptionsOverrideHeader {
   val HEADER_KEY = "X-Frame-Options-Override"
+}
+class HandleXFrameOptionsOverrideHeader @Inject()(implicit  val mat: Materializer) extends Filter {
 
   def apply(nextFilter: RequestHeader => Future[Result])(requestHeader: RequestHeader): Future[Result] = {
     nextFilter(requestHeader).map { result =>
-      result.header.headers.get(HEADER_KEY).fold(result) { value =>
+      result.header.headers.get(HandleXFrameOptionsOverrideHeader.HEADER_KEY).fold(result) { value =>
         result.withHeaders("X-Frame-Options" -> value)
       }
     }
