+++ app/views/staff/cas.scala.html
@@ -46,7 +46,7 @@
                     <div class="form-field">
                         <label class="label" for="cas.period">Validity period (weeks)</label>
                         <select required name="cas.period" >
-                        @for(w <- Range(1, 53)) {
+                        @for(w <- 1 until 53) {
                             <option value="@w">@w</option>
                         })
                         </select>
